sepearte into different classes.
